refactor: move common api types to `common::api`
