+++ mobile/native/src/lib.rs
@@ -2,6 +2,7 @@ mod api;
 pub mod api_model;
 mod bridge_generated;
 pub mod calculations;
+pub mod common;
 pub mod event;
 pub mod ln_dlc;
 pub mod logger;
